server/requirements.txt
astroid==2.1.0
awsebcli==3.14.8
blessed==1.15.0
botocore==1.12.68
cached-property==1.5.1
cachetools==3.0.0
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
Django==2.1.4
django-cors-headers==2.4.0
django-environ==0.4.5
django-extensions==2.1.4
django-rest-swagger==2.2.0
djangorestframework==3.9.0
docker==3.6.0
docker-compose==1.21.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
google-auth==1.6.2
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
pathspec==0.5.5
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.18.4
rsa==4.0
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.54.0
wrapt==1.10.11
